% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBrowser.R
\name{arrangeInsertion}
\alias{arrangeInsertion}
\title{Arrange insertions in mulitple lines.}
\usage{
arrangeInsertion(query, insertion)
}
\arguments{
\item{query}{The gapped query line among the two-line alignment representation.}

\item{insertion}{A vector containing 1-based insertion start position in query. The insertion lengths are stored as a vector in the match.length attribution.}
}
\value{
A HTML string tagList containing lines of insertions. Each line starts with ##########.
}
\description{
Try to put insertion sequences with the upstream end aligned to the insertion position. If the spans of two insertion sequences overlap, then put the downstream one to another line.
}
\examples{
query <- "----------------------------TTGGCCTGGAGTAAAAGCATGAT----------GATCGGAATGATTACAGGTAAA------------------------------------------------------------------------------CAAAAAA"
refLine <- "aGTTGGCTAGTCAATACCTGAAGAGAGATTGGCCTGGAGTAAAAGC-TGAtaAAAGCTGATGATCGGAATGATTACAGGTAAATTAGTAGTTTTTGCCTATTTTCTTTAGAAACGGTTTTACTTAAAGCTATGTTACATATAGATAATGTAACACTCTAGt-------"
insertion <- refLine |> gregexpr(pattern = "-+") |> _[[1]]
insertionLines <- arrangeInsertion(query, insertion)
}
