% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{calInsertionCount}
\alias{calInsertionCount}
\title{Get insertion frequencies}
\usage{
calInsertionCount(refList, cuts, maxCutDown)
}
\arguments{
\item{refList}{gapped reference lines in the two-line alignments}

\item{cuts}{cut sites}

\item{maxCutDown}{the maximal length of reference part downstream to the cut site}
}
\value{
tibble with columns count (insertion frequencies) and pos (relative position to cut site)
}
\description{
Get positional insertion frequencies relative to the cut site.
}
\examples{
insertCount <- calInsertionCount(refList, cuts, maxCutDown)
}
