% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{continuousDistribution}
\alias{continuousDistribution}
\title{Plot weighted distribution of continuous variable}
\usage{
continuousDistribution(algTibble, distriTempFile)
}
\arguments{
\item{algTibble}{selected tibble containing only columns to fit density and the count column}

\item{distriTempFile}{the pdf tempfile}
}
\value{
the html iframe tag to be rendered in shiny::renderUI
}
\description{
Visualize the distribution of continuous variable through density plot weighted by query counts.
}
\examples{
iframeTag <- continuousDistribution(algTibble, distriTempFile)
}
