% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseSubstitute.R
\name{countBaseSubstitute}
\alias{countBaseSubstitute}
\title{Count base substitution of query in reference.}
\usage{
countBaseSubstitute(algTibble)
}
\arguments{
\item{algTibble}{tibble containing gaggped reference and query in the two-line alignment.}
}
\value{
tibble with sub labels like X>X and corresponding counts.
}
\description{
Mutation is like "A>C". Insertion is like "->A". Deletion is like "A>-".
}
\examples{
tb <- countBaseSubstitute(algTibble)
}
