% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{discreteDistribution}
\alias{discreteDistribution}
\title{Plot weighted distribution of discrete variable}
\usage{
discreteDistribution(algTibble, distriTempFile)
}
\arguments{
\item{algTibble}{selected tibble containing only columns to aggregate and the count column}

\item{distriTempFile}{the pdf tempfile}
}
\value{
the html iframe tag to be rendered in shiny::renderUI
}
\description{
Visualize the distribution of discrete variable through bar plot weighted by query counts.
}
\examples{
iframeTag <- discreteDistribution(algTibble, distriTempFile)
}
