% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microHomology.R
\name{drawMicroHomologyHeatmap}
\alias{drawMicroHomologyHeatmap}
\title{Visualize micro-homology}
\usage{
drawMicroHomologyHeatmap(
  mhTibbleSub,
  refEnd1Start2TibbleMicro,
  maxCut1,
  maxCut2,
  maxCut1down,
  maxCut2down,
  mode,
  mhMatrixTempFile
)
}
\arguments{
\item{mhTibbleSub}{the tibble recording micro-homology ranges in rows}

\item{refEnd1Start2TibbleMicro}{the tibble containing actual counts of editing output for a certain reference}

\item{maxCut1}{the maximal length of the ref1 part upstream to the cut site}

\item{maxCut2}{the maximal length of the ref2 part upstream to the cut site}

\item{maxCut1down}{the maximal length of the ref1 part downstream to the cut site}

\item{maxCut2down}{the maximal length of the ref2 part downstream to the cut site}

\item{mode}{if set to "separate", then meaningly distribute the actual counts of editing output to the all micro-homology equivalence}

\item{mhMatrixTempFile}{the pdf tempfile to save ggplot result}
}
\value{
the html iframe tag to be displayed by shiny::renderUI
}
\description{
Visualize the actual counts of editing output by heatmap and micro-homologies by paths passing the whole mhs.
}
\examples{
iframeTag <- drawMicroHomologyHeatmap(mhTibbleSub, refEnd1Start2TibbleMicro, maxCut1, maxCut2, maxCut1down, maxCut2down, mode, mhMatrixTempFile)
}
