% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{drawPositionalLogo}
\alias{drawPositionalLogo}
\title{Use ggseqlogo to represent frequency matrix}
\usage{
drawPositionalLogo(baseFreq, method, namespace, refTempFile)
}
\arguments{
\item{baseFreq}{frequency matrix of base}

\item{method}{ggseqlogo method ("bits" or "probability") influence the heights of the logo}

\item{namespace}{alphabet of logo}

\item{refTempFile}{temp pdf file to save ggplot output}
}
\value{
html iframe tag to be rendered by shiny::renderUI
}
\description{
Base frequencies (including deletion frequencies expressed in "-") at each position relative to cut site are visualized by ggseqlogo.
}
\examples{
iframeTag <- drawPositionalLogo(baseFreq, "bits", "ACGT", refTempFile)
}
