% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{drawPositionalReads}
\alias{drawPositionalReads}
\title{Draw resolution tibble}
\usage{
drawPositionalReads(tibb, maxCut, maxCutdown, refTempFile)
}
\arguments{
\item{tibb}{resolution tibble to plot}

\item{maxCut}{the maximal length of reference part upstream to cut site}

\item{refTempFile}{temp pdf file to save ggplot output}

\item{maxCutDown}{the maximal length of reference part downstream to cut site}
}
\value{
html iframe tag to be rendered by shiny::renderUI
}
\description{
Resolution tibble from matrix adapted to specified resolution are visualized to ggplot2.
}
\examples{
iframeTag <- drawPositionalReads(tibb, maxCut, maxCutdown, refTempFile)
}
