% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{drawPositionalStatic}
\alias{drawPositionalStatic}
\title{Stack bar plot of positional base frequencies or deletion/snp/match frequencies}
\usage{
drawPositionalStatic(inputTibble, insertCount, refTempFile)
}
\arguments{
\item{inputTibble}{tibble with columns count (base frequencies), pos (relative position to cut site) and type (base type)}

\item{insertCount}{tibble with columns count (insertion frequencies) and pos (relative position to cut site)}

\item{refTempFile}{temp pdf file to save ggplot output}
}
\value{
html iframe tag to be rendered by shiny::renderUI
}
\description{
Base frequencies (including deletion frequencies expressed in "-") or deletion/snp/match frequencies at each position relative to cut site are visualized as stacked bar plot. Insertions are expressed by step plot with step height as the insertion frequencies and step position centered at where the insertions actually happens (between two base positions).
}
\examples{
iframeTag <- drawPositionalStatic(baseTibble, insertCount, tempFile)
iframeTag <- drawPositionalStatic(MSDTibble, insertCount, tempFile)
}
