% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{extendToAlignCut}
\alias{extendToAlignCut}
\title{Align sequence by cut site and pad on both sides}
\usage{
extendToAlignCut(seqs, cuts)
}
\arguments{
\item{seqs}{sequence vector}

\item{cuts}{cut site vector}
}
\value{
matrix with each row a padded sequence
}
\description{
All sequences are aligned according to the cut site. Then each sequence is padded by "-" on both sides such that the results sequences are of the same length.
}
\examples{
refMat <- extendToAlignCut(seqs)
}
