% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{extendToSameLength}
\alias{extendToSameLength}
\title{Right pad sequences}
\usage{
extendToSameLength(seqs)
}
\arguments{
\item{seqs}{sequence vector}
}
\value{
matrix with each row a right-padded sequence
}
\description{
Right pad sequences to the same length as the longest sequence by "-".
}
\examples{
refGapMat <- extendToSameLength(seqs)
}
