% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcDeletion.R
\name{getArcDelTibble}
\alias{getArcDelTibble}
\title{Get deletion info}
\usage{
getArcDelTibble(algTibble)
}
\arguments{
\item{algTibble}{tibble with alignment info}
}
\value{
tibble with deletion info
}
\description{
Get the range of deletions for each query.
}
\examples{
arcDelTibble <- getArcDelTibble(algTibble)
}
