% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcDeletion.R
\name{getArcSegment}
\alias{getArcSegment}
\title{Express deletions by arcs}
\usage{
getArcSegment(arcDelTibble, segNum)
}
\arguments{
\item{arcDelTibble}{tibble containing deletion info}

\item{segNum}{segment number to smooth the arc}
}
\value{
tibble with arc segments
}
\description{
Infer the arcs segments from insertion info.
}
\examples{
arcDelSegTibble <- getArcSegment(arcDelTibble, segNum)
}
