% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicClassify.R
\name{getIndelTypes}
\alias{getIndelTypes}
\title{Classify editing output}
\usage{
getIndelTypes(algTibble)
}
\arguments{
\item{algTibble}{a tibble containing editing information}
}
\value{
a summary tibble with the column indelType ("WT", "deletion", "insertion", "indel") and count (the count of indelType)
}
\description{
Classify the CRISPR editing output based on whether there is deletion and insertion.
}
\examples{
indelTypeTibble <- getIndelTypes(algTibble)
}
