% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicClassify.R
\name{getIndelTypesEx}
\alias{getIndelTypesEx}
\title{Classify editing output with discrimination of templated and random insertions}
\usage{
getIndelTypesEx(algTibble)
}
\arguments{
\item{algTibble}{a tibble containing editing information}
}
\value{
a summary tibble with the column indelType ("WT", "deletion", "templated", "random", "temprand", "tempdel", "randdel", "full") and count (the count of indelType)
}
\description{
Classify the CRISPR editing output based on whether there is deletion, templated insertion and random insertion.
}
\examples{
indelTypeTibbleEx <- getIndelTypesEx(algTibble)
}
