% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBrowser.R
\name{getMarkdownFromAlign}
\alias{getMarkdownFromAlign}
\title{Transform two-line alignments to html details tags.}
\usage{
getMarkdownFromAlign(algTibble)
}
\arguments{
\item{algTibble}{A tibble containing information of two-line alignments. algTibble includes reference id and sequence, cleavage site, query count, gapped reference and query sequences.}
}
\value{
A list alternatively contains reference and query line html raw strings.
}
\description{
Render two-line alignments in html format. Support folding insertions, highlighting snps, coloring extended parts in references and annote reference id and query count.
}
\examples{
allMd <- getMarkdownFromAlign(algTibble)
}
