% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microHomology.R
\name{getMicroHomologyTibble}
\alias{getMicroHomologyTibble}
\title{Get range tibble of micro-homology}
\usage{
getMicroHomologyTibble(ref1, ref2, cut1, cut2)
}
\arguments{
\item{ref1}{the reference upstream to the cut site}

\item{ref2}{the reference downstream to the cut site}

\item{cut1}{the cut site in ref1}

\item{cut2}{the cut site in ref2}
}
\value{
a tibble with the four columns: pos1low (the row start position of mh), pos1up (the row end position of mh), shift (pos1up - pos2up), cls (the index of mh)
}
\description{
Record the ramge pf micro-homology in rows of tibble.
}
\examples{
mhTibble <- getMicroHomologyTibble(ref1, ref2, cut1, cut2)
}
