% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonInsertion.R
\name{getPolyInsTibble}
\alias{getPolyInsTibble}
\title{Get insertion info}
\usage{
getPolyInsTibble(algTibble)
}
\arguments{
\item{algTibble}{tibble with alignment info}
}
\value{
tibble with insertion info
}
\description{
Get the position and length of insertions for each query.
}
\examples{
polyInsTibble <- getPolyInsTibble(algTibble)
}
