% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonInsertion.R
\name{getPolyXY}
\alias{getPolyXY}
\title{Express insertions by triangle}
\usage{
getPolyXY(polyInsTibble, mode)
}
\arguments{
\item{polyInsTibble}{tibble containing insertion info}

\item{mode}{"up" or "down", indicate the triangle direction}
}
\value{
tibble with axes of triangle vertex
}
\description{
Infer the axes of triangle vertex from insertion info.
}
\examples{
polyXY <- getPolyXY(polyInsTibble, mode)
}
