% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{getPositionalBaseFreq}
\alias{getPositionalBaseFreq}
\title{Get query base frequencies}
\usage{
getPositionalBaseFreq(queryMat, counts)
}
\arguments{
\item{queryMat}{a matrix of query lines}

\item{counts}{the counts for lines in queryMat}
}
\value{
a matrix with five rows named "-", "A", "C", "G", "T". Each row records the positional frequency of each base.
}
\description{
The counts are grouped and summed with each column of queryMat as labels. The insertions (relative to reference line) in query lines are collapsed. Then the query lines are aligned according to cut site.
}
\examples{
baseFreq <- getPositionalBaseFreq(queryMat, counts)
}
