% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{getPositionalMSDTibble}
\alias{getPositionalMSDTibble}
\title{Get query mutation frequencies}
\usage{
getPositionalMSDTibble(queryMat, refMat, counts, cut)
}
\arguments{
\item{queryMat}{a matrix of query lines}

\item{refMat}{a matrix of reference lines}

\item{counts}{the counts for lines in queryMat}

\item{cut}{the cut site}
}
\value{
a tibble with columns count (frequencies of mutation), pos (relative position to cut site) and type (one of "delete", "SNP", "match").
}
\description{
The counts are grouped and summed with each column of queryMat as labels. The insertions (relative to reference line) in query lines are collapsed. Then the query lines are aligned according to cut site.
}
\examples{
MSDTibble <- getPositionalMSDTibble(queryMat, refMat, counts, maxCut)
}
