% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{getPositionalReads}
\alias{getPositionalReads}
\title{Display values stored in matrix rows}
\usage{
getPositionalReads(mat, counts, resolution, cut)
}
\arguments{
\item{mat}{matrix to process}

\item{counts}{repeats of matrix rows}

\item{resolution}{display resolution to adapted to}

\item{cut}{cut site}
}
\value{
tibble with columns x (resolution matrix column index), y (resolution matrix row index), value (resolution matrix value) and height (resolution matrix pixel heights for rows)
}
\description{
Matrix rows with repeats are adapted to specified resolution. Then the matrix is melted for ploting by ggplot.
}
\examples{
resolutionTibble <- getPositionalReads(mat, counts, resolution, cut)
}
