% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microHomology.R
\name{getRefEnd1Start2Tibble}
\alias{getRefEnd1Start2Tibble}
\title{Get editing output counts}
\usage{
getRefEnd1Start2Tibble(algTibble, microRefId)
}
\arguments{
\item{algTibble}{the tibble with columns ref1End (the reference end position of the first alignment block), cut1 (the cut site in ref1), ref2Start (the reference start position of the second alignment block), cut2 (the cut site in ref2), refId (the reference id), count (the count of the query)}

\item{microRefId}{the reference id to harvest}
}
\value{
the tibble with the columns pos1 (the end position of the upstream alignment block relative to the cut site), pos2 (the end position of the downstream alignment block relative to the cut site), count (the count of the query), shift (pos1 - pos2)
}
\description{
Get the actual counts of editing output for a specific reference.
}
\examples{
refEnd1Start2Tibble <- getRefEnd1Start2Tibble(algTibble, microRefId)
}
