% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microHomology.R
\name{getRefEnd1Start2TibbleMicro}
\alias{getRefEnd1Start2TibbleMicro}
\title{Duplicate the editing output count upon micro-homology equivalence}
\usage{
getRefEnd1Start2TibbleMicro(refEnd1Start2Tibble, mhTibbleSub)
}
\arguments{
\item{refEnd1Start2Tibble}{the tibble recording the counts of editing outputs}

\item{mhTibbleSub}{the tibble recording micro-homology ranges in rows}
}
\value{
the tibble duplicates the counts of editing outputs upon micro-homology equivalence
}
\description{
For each editing output, if it is located in a micro-homology, then duplicate its count to all equivalence editing output in the micro-homology range.
}
\examples{
refEnd1Start2TibbleMicro <- getRefEnd1Start2TibbleMicro(refEnd1Start2Tibble, mhTibbleSub)
}
