% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicClassify.R
\name{indelTypePiePlot}
\alias{indelTypePiePlot}
\title{Pie plot of editing classfication}
\usage{
indelTypePiePlot(indelTypeTibble, classifyTempFile)
}
\arguments{
\item{indelTypeTibble}{the tibble with classification and counts}

\item{classifyTempFile}{the pdf tempfile}
}
\value{
the html iframe to be displayed by shiny:renderUI
}
\description{
Visualize the classification of editing output by pie plot.
}
\examples{
iframeTag <- indelTypePiePlot(indelTypeTibble, classifyTempFile)
}
