% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwisePlot}
\alias{pairwisePlot}
\title{Pairwise plot for two columns}
\usage{
pairwisePlot(pairTibble, xscale, yscale, method, span, pairwiseTempFile)
}
\arguments{
\item{pairTibble}{tibble with two columns to apply pairwise plot}

\item{xscale}{transform of x axis}

\item{yscale}{transform of y axis}

\item{method}{method for smooth fit}

\item{span}{span for smooth fit}

\item{pairwiseTempFile}{the pdf temp file}
}
\value{
the html iframe tag to be displayed in shiny::renderUI
}
\description{
Scatter plot with smooth fit for two columns.
}
\examples{
# example code
}
