% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcDeletion.R
\name{plotArcDelTibble}
\alias{plotArcDelTibble}
\title{Plot deletion arcs}
\usage{
plotArcDelTibble(arcDelSegTibble, limits, arcDeleteTempFile)
}
\arguments{
\item{arcDelSegTibble}{tibble with arc segments}

\item{limits}{the limits of the reference coordinate}

\item{arcDeleteTempFile}{the pdf temp file}
}
\value{
the html iframe tag to be displayed in shiny::renderUI
}
\description{
Visualize deletions by arcs. The arc line width is the deletion count. The arc range is the deletion range.
}
\examples{
iframeTag <- plotArcDelTibble(arcDelSegTibble, limits, arcDeleteTempFile)
}
