% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerFrequencies.R
\name{plotKmerFrequencies}
\alias{plotKmerFrequencies}
\title{Kmer enrichment of target}
\usage{
plotKmerFrequencies(kmerTibble, kmerPdfTempFile)
}
\arguments{
\item{kmerTibble}{tibble with kmer count for target and nontarget editing output}

\item{kmerPdfTempFile}{the pdf temp file}
}
\value{
the html iframe tag to be displayed in shiny::renderUI
}
\description{
Stack the target and nontarget editing output count for each kmer from sgRNA.
}
\examples{
iframeTag <- plotKmerFrequencies(kmerTibble, kmerPdfTempFile)
}
