% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonInsertion.R
\name{plotPolyInsTibble}
\alias{plotPolyInsTibble}
\title{Plot insertion triangles}
\usage{
plotPolyInsTibble(polyXY, limits, polyInsertTempFile)
}
\arguments{
\item{polyXY}{tibble with axes of triangle vertex}

\item{limits}{the limits of the reference coordinate}

\item{polyInsertTempFile}{the pdf temp file}
}
\value{
the html iframe tag to be displayed in shiny::renderUI
}
\description{
Visualize insertions by triangles. The triangle height is the insertion count. The triangle width is the insertion size.
}
\examples{
iframeTag <- plotPolyInsTibble(polyXY, limits, polyInsertTempFile)
}
