% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{posMatrixToTibble}
\alias{posMatrixToTibble}
\title{Melt matrix}
\usage{
posMatrixToTibble(mat, cut)
}
\arguments{
\item{mat}{the base frequency matrix to melt}

\item{cut}{the cut site}
}
\value{
melted matrix as tibble with columns count, pos and type
}
\description{
Melt the matrix with position relative to cut site recorded in column pos and base type (rownames) recorded in column type. matrix values are recorded in column count.
}
\examples{
baseTibble <- posMatrixToTibble(baseFreq, maxCut)
}
