% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{resolutionMatrix}
\alias{resolutionMatrix}
\title{Adapter matrix to display resolution}
\usage{
resolutionMatrix(mat, counts, resolution)
}
\arguments{
\item{mat}{the matrix to display}

\item{counts}{the repeats of matrix rows}

\item{resolution}{the output matrix row number to fit the display resolution}
}
\value{
description of the return value
}
\description{
Because display resolution is only 1080 in height, matrix with large number of rows cannot be displayed smoothly and correctly. This function transforms matrix with large row numbers to a display matrix with specified row number (resolution). Each row of the matrix may repeat multiple times (counts).
}
\examples{
resolutionMatrix <- resolutionMatrix(mat, counts, resolution)
}
