% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBrowser.R
\name{snpHighlight}
\alias{snpHighlight}
\title{Highlight SNP.}
\usage{
snpHighlight(refSeg, querySeg)
}
\arguments{
\item{refSeg}{Reference segment.}

\item{querySeg}{Query segment.}
}
\value{
A tagList with mismatch surrounded in highlight \if{html}{\out{<span>}} tags.
}
\description{
Highlight mismatch between reference and query sequences.
}
\examples{
refSeg <- "aGTTGGCTAGTCAATACCTGAAGAGAGATTGGCCTGGAGTAAAAG"
querySeg <- "----------------------------TTGGCCTGGAGTAAAAG"
mdHigh <- snpHighlight(refSeg, querySeg)
}
