% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalStatistics.R
\name{vectorToStringVector}
\alias{vectorToStringVector}
\title{Split character vector according to lengths}
\usage{
vectorToStringVector(vec, strLens)
}
\arguments{
\item{vec}{a character vector}

\item{strLens}{length of each substring}
}
\value{
a vector of substrings
}
\description{
Split a character vector into consecutive subvectors with specified length, and transform character subvectors to substrings.
}
\examples{
vectorToStringVector(c("A", "B", "C", "D", "E"), c(2, 3))
[1] "AB"  "CDE"
}
